requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
httpx==0.25.2
pydantic==2.5.0
pydantic-settings==2.1.0
python-multipart==0.0.6
PyMuPDF==1.26.6
Pillow>=10.4
python-dotenv==1.0.0
opencv-python
numpy
pandas
pdfplumber
camelot-py
tabula-py
python-magic
fidel==0.1.0

